+++ tests/__init__.py
@@ -10,3 +10,7 @@ mac_arm64 = sys.platform == "darwin" and platform.machine() == "arm64"
 mac_arm64_unsupported = pytest.mark.skipif(
     mac_arm64, reason="can't be tested under mac arm64 due to h5py incompatibility"
 )
+mac_only = pytest.mark.skipif(
+    sys.platform.startswith("win") or sys.platform.startswith("lin"),
+    reason="can't be tested under windows due to eccodes incompatibility",
+)